Reorganize strmatcher's package structure
