+++ common/strmatcher/indexmatcher_mph.go
