+++ common/strmatcher/matchergroup_simple.go
@@ -1,0 +1,6 @@
+package strmatcher
+
+type matcherEntry struct {
+	m  Matcher
+	id uint32
+}
